1.新增标签管理

+++ blog/src/main/java/top/yueqiao/blog/service/impl/CategoryServiceImpl.java
@@ -5,10 +5,14 @@ import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 import top.yueqiao.blog.domain.PageQuery;
 import top.yueqiao.blog.domain.PageResult;
+import top.yueqiao.blog.domain.entity.Blog;
 import top.yueqiao.blog.domain.entity.Category;
+import top.yueqiao.blog.exception.ServiceException;
+import top.yueqiao.blog.mapper.BlogMapper;
 import top.yueqiao.blog.mapper.CategoryMapper;
 import top.yueqiao.blog.service.ICategoryService;
 
@@ -19,6 +23,8 @@ import top.yueqiao.blog.service.ICategoryService;
 @Service
 public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> implements ICategoryService {
 
+    @Resource
+    private BlogMapper blogMapper;
 
     @Override
     public PageResult<Category> selectPageCategoryList(Category category, PageQuery pageQuery) {
@@ -45,6 +51,11 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
 
     @Override
     public int deleteCategoryByIds(Integer[] ids) {
+        LambdaQueryWrapper<Blog> lqw = new LambdaQueryWrapper<Blog>()
+                .eq(Blog::getCategoryId, CollUtil.toList(ids));
+        if (blogMapper.selectCount(lqw) > 0) {
+            throw new ServiceException("分类下存在博客,无法删除");
+        }
         return baseMapper.deleteByIds(CollUtil.toList(ids));
     }
